refactor: move build_search_strategy from cli/commands to limen.yaml.compiler
