+++ limen/yaml/compiler.py
@@ -1,9 +1,12 @@
 from datetime import date
 from typing import Any
 
+from limen.experiment import GridStrategy
+from limen.experiment import RandomStrategy
 from limen.experiment.manifest_core import MLManifest
 from limen.experiment.manifest_core import Manifest
 from limen.experiment.manifest_core import RuleBasedManifest
+from limen.experiment.param_domain import ParamDomain
 from limen.yaml.errors import ResolutionError
 from limen.yaml.resolver import resolve
 
@@ -271,3 +274,33 @@ class CompiledSFD:
         if self._manifest_cache is None:
             self._manifest_cache = build_manifest(self._yaml)
         return self._manifest_cache
+
+
+def build_search_strategy(yaml_dict: dict[str, Any]) -> RandomStrategy | GridStrategy:
+
+    '''
+    Build a search strategy from a validated YAML experiment dict.
+
+    Args:
+        yaml_dict (dict): Parsed and validated YAML experiment dict
+
+    Returns:
+        RandomStrategy | GridStrategy: Configured search strategy
+
+    Raises:
+        ValueError: If uel.search_strategy is not a mapping
+
+    '''
+
+    uel_cfg = yaml_dict.get('uel', {})
+    sfd_cfg = yaml_dict.get('sfd', {})
+    strategy_cfg = uel_cfg.get('search_strategy', {})
+    if not isinstance(strategy_cfg, dict):
+        raise ValueError(
+            f"'uel.search_strategy' must be a mapping, got {type(strategy_cfg).__name__}"
+        )
+    params = {k: list(v) for k, v in (sfd_cfg.get('params') or {}).items()}
+    domain = ParamDomain(params)
+    if strategy_cfg.get('type', 'random') == 'grid':
+        return GridStrategy(domain)
+    return RandomStrategy(domain)
